feat: Add email notification for email address changes
Added an email notification system for when a user's email address is changed. This includes a new dynamic email updated template which provides different content for security alerts (to the old email) and confirmations (to the new email). Also, an updated mail utility with a send email change notification function to dispatch these conditional emails.