feat: Implement minimap for upload preview modal
Introduced a minimap feature to the upload preview modal for better navigation on zoomed-in images. Created a new upload preview minimap component to render the view context. The upload preview modal was updated to track pan/zoom state via the  callback from react-zoom-pan-pinch and pass this data to the minimap. The minimap is displayed on desktop views to show the current viewport.