feat: Improve navigation and context in authenticated layout
Introduces several enhancements to the main authenticated layout. The breadcrumb navigation now displays the case name for improved clarity on dynamic pages. In the sidebar, the 'History' menu has been renamed to 'Results', and active link highlighting is more accurate for nested routes. And the header has been optimized to be responsive on smaller devices and visually separated with a new drop shadow.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>