feat: Implement recalculation poller hook and state management
Added client-side logic for the case recalculation feature. Created the use recalculation poller custom hook which uses TanStack Query to poll the get recalculation status action and executes a callback on completion. Also updated the results store to track cases with pending recalculations to enable the interface to trigger and monitor this process.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>