feat: Implement the interface for case history tab
Created the front-end components to display the case audit trail in the history tab. This includes case history log which acts as the main component that fetches data using the use case history hook hook and handles loading/error/empty states. The history log timeline which is a component to structure the log entries into a day-by-day timeline view. The history log timeline event component which renders a single batch of changes, including user details, timestamps, and a formatted before/after view of the data.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>