refactor: Generalize location dropdown component for broader reusability
Refactored the location dropdown component to be a highly reusable, generic form input. Switched from form field to use controller to decouple it from a specific form context. Added a wide range of new props for customizing styling and behavior The cascading reset logic was also improved. This makes the component suitable for use in various forms across the application.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>