fix: Correctly handle two-factor for social providers
Addressed a bug where users with two-factor authentication (2FA) enabled were not being prompted for a 2FA code when signing in via an social provider. The fix includes updated to create a temporary auth session for OAuth users who have 2FA enabled, and a new callback was implemented to check for this temporary session after any sign-in. If the session exists, it intercepts the redirect and forces the user to the two-factor sign-in page to complete authentication. Temporary session now also includes the provider name.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>