feat: Implement navigation logic for edit image modal
Activated the proceed button in the edit image modal by implementing the navigation logic. The component now uses Next.js' `useRouter` to navigate to the a dedicated image editing page. It handles the user's choice to open the editor in either the current tab (`router.push`) or a new tab (`window.open`).

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>