feat: Implement interactive image display for editor
Created the editor image display component to serve as the main interactive canvas for the image editor. This component uses `react-zoom-pan-pinch` to enable panning and zooming of the image. It uses the use rendered image hook to accurately calculate image dimensions and correctly overlays detection data using the editor bounding box component. A loading state is provided while the image is being fetched.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>