feat: Implement minimap for image editor
Created an editor image minimap component. This component provides a scaled-down preview of the main image and displays a dynamic viewport rectangle that represents the user's current zoom level and pan position. It consumes state from the `react-zoom-pan-pinch` wrapper to calculate the viewport's size and position that offers users better spatial awareness when navigating large images.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>