feat: Add functional minimap toggle to editor
Added a functional, toggleable minimap to the annotation editor. State for minimap visibility and the image's transform data has been lifted to the parent editor layout. The editor toolbar now includes a working button to toggle the minimap. The editor image display reports its transform state up to the layout, which is then passed to the conditionally rendered editor image minimap component to display the correct viewport.

Co-authored-by: Precious Andal <168273426+cinnamongirl29@users.noreply.github.com>